oops wrote float instead of double

+++ incs_bonus/elem.h
@@ -6,7 +6,7 @@
 /*   By: nseniak <nseniak@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/12/20 17:52:21 by ngiroux           #+#    #+#             */
-/*   Updated: 2023/01/18 12:14:54 by nseniak          ###   ########.fr       */
+/*   Updated: 2023/01/18 13:09:15 by nseniak          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -40,7 +40,7 @@ typedef struct s_sphere
 	t_vect	pos;
 	double	radius;
 	t_rgb	rgb;
-	float	n;
+	double	n;
 }	t_sphere;
 
 typedef struct s_plane
@@ -48,7 +48,7 @@ typedef struct s_plane
 	t_vect	pos;
 	t_vect	dir;
 	t_rgb	rgb;
-	float	n;
+	double	n;
 }	t_plane;
 
 typedef struct s_cylinder
@@ -59,7 +59,7 @@ typedef struct s_cylinder
 	double	radius;
 	double	height;
 	t_rgb	rgb;
-	float	n;
+	double	n;
 }	t_cylinder;
 
 #endif /* !ELEM_H */
